add ollama filename structure
